Consolidate event listeners into WorldListener class
