--- src/main/java/net/thenextlvl/utilities/listener/ExplosionListener.java
@@ -1,36 +1,0 @@
-/*
- * Builder's Utilities is a collection of a lot of tiny features that help with building.
- * Copyright (C) Arcaniax-Development
- * Copyright (C) Arcaniax team and contributors
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <https://www.gnu.org/licenses/>.
- */
-package net.thenextlvl.utilities.listener;
-
-import lombok.RequiredArgsConstructor;
-import net.thenextlvl.utilities.UtilitiesPlugin;
-import org.bukkit.event.EventHandler;
-import org.bukkit.event.EventPriority;
-import org.bukkit.event.Listener;
-import org.bukkit.event.entity.EntityExplodeEvent;
-
-@RequiredArgsConstructor
-public class ExplosionListener implements Listener {
-    private final UtilitiesPlugin plugin;
-
-    @EventHandler(priority = EventPriority.HIGHEST, ignoreCancelled = true)
-    public void onExplode(EntityExplodeEvent event) {
-        event.setCancelled(plugin.config().disableExplosions());
-    }
-}
